requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
qdrant-client>=1.10.0
httpx>=0.24.1
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
numpy>=1.24.3
langgraph>=0.0.21